Implemented Edge To Edge
