+++ app/build.gradle.kts
@@ -67,7 +67,6 @@ dependencies {
     implementation("androidx.compose.material3:material3")
     implementation("androidx.compose.material:material-icons-extended")
 
-
     //Navigation
     implementation("androidx.navigation:navigation-compose:2.7.5")
