+++ app/src/main/AndroidManifest.xml
@@ -5,6 +5,7 @@
     <application
         android:name=".FindOutApp"
         android:allowBackup="true"
+
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
