+++ app/src/main/java/com/leoapps/findout/creation/form/presentation/composables/TopBar.kt
@@ -1,6 +1,6 @@
 package com.leoapps.findout.creation.form.presentation.composables
 
-import androidx.compose.animation.animateColorAsState
+import androidx.compose.animation.core.animateDpAsState
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -9,6 +9,7 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.statusBarsPadding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material.ripple.rememberRipple
@@ -20,7 +21,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.unit.dp
 import com.leoapps.findout.creation.form.presentation.model.FormCreationUiAction
 
@@ -30,16 +31,18 @@ internal fun TopBar(
     isContentScrolled: Boolean,
     onAction: (FormCreationUiAction) -> Unit
 ) {
-    val contentColor by animateColorAsState(
-        targetValue = if (isContentScrolled) Color.Gray else Color.White,
+    val elevation by animateDpAsState(
+        targetValue = if (isContentScrolled) 8.dp else 0.dp,
         animationSpec = tween(),
         label = "",
     )
 
     Box(
         modifier = Modifier
+            .statusBarsPadding()
             .fillMaxWidth()
-            .background(contentColor)
+            .shadow(elevation)
+            .background(MaterialTheme.colorScheme.surface)
             .padding(16.dp)
     ) {
         Icon(
